feat: rename /points route to /rewards with backward compat redirect
- Move src/app/(mobile-ui)/points/ → src/app/(mobile-ui)/rewards/
- Create /points and /points/invites redirect pages for backward compat
- Update all href and router.push references to /rewards
- Keeps /points working for existing deep links, push notifications, bookmarks
