+++ src/app/(mobile-ui)/rewards/invites/page.tsx
@@ -1,0 +1,148 @@
+'use client'
+
+import PageContainer from '@/components/0_Bruddle/PageContainer'
+import Card from '@/components/Global/Card'
+import { getCardPosition } from '@/components/Global/Card/card.utils'
+import NavHeader from '@/components/Global/NavHeader'
+import PeanutLoading from '@/components/Global/PeanutLoading'
+import TransactionAvatarBadge from '@/components/TransactionDetails/TransactionAvatarBadge'
+import { VerifiedUserLabel } from '@/components/UserHeader'
+import { useAuth } from '@/context/authContext'
+import { invitesApi } from '@/services/invites'
+import { useQuery } from '@tanstack/react-query'
+import { useRouter } from 'next/navigation'
+import { STAR_STRAIGHT_ICON } from '@/assets'
+import Image from 'next/image'
+import EmptyState from '@/components/Global/EmptyStates/EmptyState'
+import { getInitialsFromName } from '@/utils/general.utils'
+import { type PointsInvite } from '@/services/services.types'
+import { formatPoints } from '@/utils/format.utils'
+import { useCountUp } from '@/hooks/useCountUp'
+import { useInView } from 'framer-motion'
+import { useRef } from 'react'
+import InviteePointsBadge from '@/components/Points/InviteePointsBadge'
+
+const InvitesPage = () => {
+    const router = useRouter()
+    const { user } = useAuth()
+    const listRef = useRef(null)
+    const listInView = useInView(listRef, { once: true, margin: '-50px' })
+
+    const {
+        data: invites,
+        isLoading,
+        isError,
+        error,
+    } = useQuery({
+        queryKey: ['invites', user?.user.userId],
+        queryFn: () => invitesApi.getInvites(),
+        enabled: !!user?.user.userId,
+    })
+
+    const totalPointsEarned =
+        invites?.invitees?.reduce((sum: number, invite: PointsInvite) => {
+            return sum + (invite.contributedPoints ?? 0)
+        }, 0) || 0
+
+    const animatedTotal = useCountUp(totalPointsEarned, {
+        storageKey: 'invites_total',
+        duration: 1.8,
+        enabled: !isLoading && !isError,
+    })
+
+    if (isLoading) {
+        return <PeanutLoading />
+    }
+
+    if (isError) {
+        console.error('Error loading invites:', error)
+        return (
+            <div className="mx-auto mt-6 w-full space-y-3 md:max-w-2xl">
+                <EmptyState icon="alert" title="Error loading invites!" description="Please contact Support." />
+            </div>
+        )
+    }
+
+    return (
+        <PageContainer className="flex flex-col">
+            <NavHeader title="Rewards" onPrev={() => router.back()} />
+
+            <section className="mx-auto mb-auto mt-10 w-full space-y-4">
+                <Card className="flex flex-col items-center justify-center gap-2 p-4">
+                    {invites?.summary?.totalLifetimeEarnedUsd !== undefined &&
+                    invites.summary.totalLifetimeEarnedUsd > 0 ? (
+                        <>
+                            <h2 className="text-center font-medium text-black">Your friends earned you</h2>
+                            <span className="text-3xl font-extrabold text-black">
+                                ${invites.summary.totalLifetimeEarnedUsd.toFixed(2)}
+                            </span>
+                            <span className="flex items-center gap-1 text-sm text-grey-1">
+                                <Image src={STAR_STRAIGHT_ICON} alt="star" width={14} height={14} />
+                                {formatPoints(totalPointsEarned)} {totalPointsEarned === 1 ? 'point' : 'points'}
+                            </span>
+                        </>
+                    ) : (
+                        <>
+                            <h2 className="text-center font-medium text-black">Your friends earned you</h2>
+                            <span className="flex items-center gap-2">
+                                <Image src={STAR_STRAIGHT_ICON} alt="star" width={20} height={20} />
+                                <span className="text-3xl font-extrabold text-black">
+                                    {formatPoints(animatedTotal)} {totalPointsEarned === 1 ? 'point' : 'points'}
+                                </span>
+                            </span>
+                        </>
+                    )}
+                </Card>
+
+                <h2 className="font-bold">People you invited</h2>
+
+                {/* Full list */}
+                <div ref={listRef}>
+                    {invites?.invitees?.map((invite: PointsInvite, i: number) => {
+                        const username = invite.username
+                        const fullName = invite.fullName
+                        const isVerified = invite.kycVerified
+                        const pointsEarned = invite.contributedPoints ?? 0
+                        // respect user's showFullName preference for avatar and display name
+                        const displayName = invite.showFullName && fullName ? fullName : username
+                        return (
+                            <Card
+                                key={invite.inviteeId}
+                                position={getCardPosition(i, invites.invitees.length)}
+                                onClick={() => router.push(`/${username}`)}
+                                className="cursor-pointer"
+                            >
+                                <div className="flex items-center justify-between gap-4">
+                                    <div className="flex items-center gap-3">
+                                        <TransactionAvatarBadge
+                                            initials={getInitialsFromName(displayName)}
+                                            userName={displayName}
+                                            isLinkTransaction={false}
+                                            transactionType={'send'}
+                                            context="card"
+                                            size="small"
+                                        />
+                                    </div>
+                                    <div className="min-w-0 flex-1 truncate font-roboto text-[16px] font-medium">
+                                        <VerifiedUserLabel
+                                            name={displayName}
+                                            username={username}
+                                            isVerified={isVerified}
+                                        />
+                                    </div>
+                                    <InviteePointsBadge
+                                        points={pointsEarned}
+                                        inView={listInView}
+                                        lifetimeEarnedUsd={invite.lifetimeEarnedUsd}
+                                    />
+                                </div>
+                            </Card>
+                        )
+                    })}
+                </div>
+            </section>
+        </PageContainer>
+    )
+}
+
+export default InvitesPage
